refactor(ui): refactor Bars
