--- ui/design/src/components/Bars/index.ts
@@ -1,4 +1,0 @@
-import { Sidebar } from './sidebar/sidebar';
-import { Topbar } from './topbar/topbar';
-
-export { Sidebar, Topbar };
